requirements.txt
# requirements.txt

# ── Core AI ──────────────────────────────
crewai==0.28.8
crewai-tools==0.1.6
llama-index==0.10.40
llama-index-vector-stores-qdrant==0.2.8
openai==1.30.1
regex>=2023.12.25,<2024.0.0

# ── Bancos de dados ──────────────────────
psycopg2-binary==2.9.9       # PostgreSQL driver
qdrant-client==1.9.1         # Qdrant client

# ── Interface e API ───────────────────────
chainlit==1.0.0
fastapi==0.100.0
uvicorn==0.23.2

# ── Observabilidade e Qualidade ───────────
langfuse==2.29.0
deepeval==0.21.45

# ── Dados e Modelagem ─────────────────────
pydantic==2.7.1
pandas==2.2.2
faker==25.0.0                # Geração de dados fake

# ── Utilitários ───────────────────────────
python-dotenv==1.0.0
httpx==0.24.1